requirements.txt
Django==1.8.4
django-rest-swagger==0.3.4
djangorestframework==3.2.3
django-filter==0.11.0
Pillow==2.9.0
mkdocs==0.14.0
pytz==2015.4
PyYAML==3.11

